feat: add `useSavesHistoryEntity` hook
